+++ IctBaden.Modbus.Test/PollServiceTests.cs
@@ -9,7 +9,7 @@ namespace IctBaden.Modbus.Test
 {
     public class PollServiceTests : IDisposable
     {
-        private static ushort _port = (SystemInfo.Platform == Platform.Windows) ? 502 : 1502; 
+        private static ushort _port = (ushort)((SystemInfo.Platform == Platform.Windows) ? 502 : 1502); 
         private readonly TestData _source;
         private ModbusMaster _master;
         private ModbusSlave _slave;
